block body creation

+++ src/burnfee.rs
@@ -2,6 +2,7 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Serialize, Deserialize, PartialEq, Clone, Debug)]
 pub struct BurnFee {
-    start: u64,
-    current: u64,
+    start: f32,
+    current: f32,
 }
+
